survey: move RecoverPayload out of rfscope (shared triage primitive)
Move the triage-grade blind demodulator (recoverPayload + sliceSymbols +
packBits) from internal/rfscope into internal/survey as exported
survey.RecoverPayload, and point the rfscope entropy analyzer at it. It already
took a survey.ClassFeatures and depended only on dsp/demod, so survey is its
natural home — and rfscope sits above survey (imports it), so the hunt survey,
which cannot import rfscope, can now reuse the same demodulator for its own
encryption triage. No behavior change; the payload tests move with it.
